modified pp_invoice_update_meta to create new entry if key does not yet exist.  Filled in gaps in save for later invoice function

+++ pp-payment/core/functions.php
@@ -133,8 +133,7 @@ function pp_invoice_create( $args, $meta = '' ) {
 	$meta_defaults = array(
 		'due_date_day' => date('j'),
 		'due_date_month' => date('n'),
-		'due_date_year' => date('Y'),
-		'sent_date' => date( 'Y-m-d' )
+		'due_date_year' => date('Y')
 		);
 
 	$args = wp_parse_args( $args, $defaults );
@@ -285,15 +284,23 @@ function pp_invoice_update_status( $invoice_id, $status ) {
 	$wpdb->query( "UPDATE ".$wpdb->payments." SET status = '$status' WHERE  id = '$invoice_id'" );
 }
 
-
+/*
+ * Updates payments meta for specified key & invoice id OR if key does not exist inserts new entry
+ *
+*/
 function pp_update_invoice_meta( $invoice_id, $meta_key, $meta_value = '' ) {
 	global $wpdb;
 
-	if( empty( $meta_value ) ) {
-		// Delete meta_key if no value is set
-		$wpdb->query( "DELETE FROM ".$wpdb->paymentsmeta." WHERE  invoice_id = '$invoice_id' AND meta_key = '$meta_key'" ); 
-	} else {
-		$wpdb->update( $wpdb->paymentsmeta, array( 'meta_value' => $meta_value ), array( 'invoice_id' => $invoice_id, 'meta_key' => $meta_key ) );
+	if ( pp_invoice_meta( $invoice_id, $meta_key ) ) { //meta key exists
+
+		if( empty( $meta_value ) ) {
+			// Delete meta_key if no value is set
+			$wpdb->query( "DELETE FROM ".$wpdb->paymentsmeta." WHERE  invoice_id = '$invoice_id' AND meta_key = '$meta_key'" ); 
+		} else {
+			$wpdb->update( $wpdb->paymentsmeta, array( 'meta_value' => $meta_value ), array( 'invoice_id' => $invoice_id, 'meta_key' => $meta_key ) );
+		}
+	} else { // meta key does not exist in paymentsmeta
+		$wpdb->insert( $wpdb->paymentsmeta, array( 'invoice_id' => $invoice_id, 'meta_key' => $meta_key, 'meta_value' => $meta_value ) );
 	}
 }
